chore(boxen): remove deprecated example and update workspace configuration
- Deleted the `box.js` example from the `examples/boxen` directory as it was no longer needed.
- Updated `pnpm-workspace.yaml` to include new benchmark directories and remove old fixture paths.
- Adjusted `pnpm-lock.yaml` to reflect the removal of dependencies related to the deleted example.
- Added new example scripts demonstrating various boxen features, including multi-color borders and gradient borders.
