--- examples/boxen/package.json
@@ -1,13 +1,0 @@
-{
-    "name": "node_example_boxen",
-    "version": "1.0.0",
-    "private": true,
-    "type": "module",
-    "scripts": {
-        "dev": "node box.js"
-    },
-    "dependencies": {
-        "@visulima/boxen": "workspace:*",
-        "@visulima/colorize": "workspace:*"
-    }
-}
