+++ packages/boxen/examples/multi-color-border-title.js
@@ -1,0 +1,27 @@
+import { boxen } from "@visulima/boxen";
+import { red, green, yellow, blue } from "@visulima/colorize";
+
+console.log("\n----------- Box with a multi color boarder and title -----------");
+
+console.log(
+    boxen("Hello, world!", {
+        headerText: "Multi Color",
+        borderColor: (border, position) => {
+            if (["top", "topLeft", "topRight"].includes(position)) {
+                return red(border);
+            }
+
+            if (position === "left") {
+                return yellow(border);
+            }
+
+            if (position === "right") {
+                return green(border);
+            }
+
+            if (["bottom", "bottomLeft", "bottomRight"].includes(position)) {
+                return blue(border);
+            }
+        },
+    }),
+);
